Support subtitle for selection lists
Support a secondary text field for to appear below the primary text of a
selection list item.

Also updated the `EdgeInsets` for the content, vertical insets changed
from 1 to 8. Otherwise the combined title and subtitle lack sufficient
padding from the item border. There is no change to the appearance of
items without a subtitle.